requirements.txt
openai
pydub 
simpleaudio
fastapi 
uvicorn 
python-dotenv 
python-multipart 
uuid
jinja2